% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_aggressive_phenotype.R
\name{determine_aggressive_phenotype}
\alias{determine_aggressive_phenotype}
\title{Determine whether patients' MS can be regarded
as aggressive.}
\usage{
determine_aggressive_phenotype(demographics, relapses, edss, eye_check = TRUE)
}
\arguments{
\item{demographics}{A tibble with basic demographic
variables.}

\item{relapses}{A tibble with relapse dates.}

\item{edss}{A tibble with EDSS scores.}

\item{eye_check}{A logical indicating whether plots
of patients classified as suffering the aggressive
disease form should be plotted for manual control.}
}
\value{
A list with a tibble containing demographic
data with added column denoting aggressive disease
($data) and text summarising the process ($text).
}
\description{
The function uses the following criteria to
determined whether patient ought to be regarded
as suffering aggressive form MS:
(i) at least 10 years of disease duration,
(ii) EDSS ≥ 6 appeared within the first 10
years of the disease,
(iii) EDSS ≥ 6 sustained for at least 6 months
after it appeared,
(iv) EDSS did not go below 6 until the end of
observation period.
Across all, the supplementary criterion needs to
be met:
(suppl.) EDSS ≥ 6 was NOT within 30 days after relapse
}
\examples{
\dontrun{
p <- here::here("data-raw", "some cool file name.xlsx")
d <- prepare_data(p)
outcome_data <- determine_aggressive_phenotype(d$id, d$relapses, d$edss, T)
}
}
