% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Read and format data for further use.}
\usage{
prepare_data(p)
}
\arguments{
\item{path}{Path to an excel file with data.}

\item{list}{A list including data file specification.}
}
\value{
A list of tibbles containing relevant
variables.
}
\description{
Using a path to the data (ideally on an external
drive to ensure patients' anonymity) and a list
containing specification of excel lists to be
imported.
}
\examples{
\dontrun{
p <- here::here("data-raw", "some cool file name.xlsx")
data <- prepare_data(p)
}
}
