% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_data_structure.R
\name{specify_data_structure}
\alias{specify_data_structure}
\title{Helper function to specify structure for data import.}
\usage{
specify_data_structure(show = TRUE)
}
\arguments{
\item{show}{A logical indicating whether the
list containing expected data structure ought
to be printed (default, TRUE) or not (FALSE).}
}
\value{
A list containing excel lists names,
labels to be used in data and column types
specification.
}
\description{
The function is supposed to be used in a wrapper
\code{prepare_data} and interrogate anytime the
source data file structure changes.
}
